+++ src/XRoadLib.Extensions.AspNet/ServiceRequestHandlerBase.cs
@@ -84,13 +84,13 @@ namespace XRoadLib.Extensions.AspNet
         protected override void HandleRequest(XRoadContext context)
         {
             if (context.HttpContext.Request.InputStream.Length == 0)
-                throw XRoadException.InvalidQuery("Empty request content");
+                throw new InvalidXRoadQueryException("Empty request content");
 
             context.Request.LoadRequest(context.HttpContext, StoragePath.GetValueOrDefault(Path.GetTempPath()), serviceManagers);
             if (context.Request.ServiceManager == null && context.Request.MetaServiceMap == null)
             {
                 var supportedProtocolsString = string.Join(", ", serviceManagers.Select(x => $@"""{x.Name}"""));
-                throw XRoadException.InvalidQuery($"Could not detect X-Road message protocol version from request message. Adapter supports following protocol versions: {supportedProtocolsString}.");
+                throw new InvalidXRoadQueryException($"Could not detect X-Road message protocol version from request message. Adapter supports following protocol versions: {supportedProtocolsString}.");
             }
 
             context.Response.Copy(context.Request);