+++ src/XRoadLib/Headers/XRoadHeader31.cs
@@ -168,7 +168,7 @@ namespace XRoadLib.Headers
                 }
             }
 
-            throw XRoadException.InvalidQuery($"Unexpected X-Road header element `{reader.GetXName()}`.");
+            throw new InvalidXRoadQueryException($"Unexpected X-Road header element `{reader.GetXName()}`.");
         }
 
         /// <inheritdoc />