+++ src/XRoadLib/Serialization/Mapping/SequenceTypeMap.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Xml;
 using XRoadLib.Extensions;
@@ -82,7 +83,8 @@ namespace XRoadLib.Serialization.Mapping
 
             while (properties.MoveNext())
             {
-                propertyMap = properties.Current;
+                propertyMap = properties.Current ?? throw new NullReferenceException();
+
                 var propertyName = propertyMap.Definition.Content.SerializedName.LocalName;
                 var propertyNode = templateNode[properties.Current.Definition.TemplateName, message.Version];
 
@@ -97,7 +99,7 @@ namespace XRoadLib.Serialization.Mapping
             }
 
             if (propertyMap == null)
-                throw XRoadException.InvalidQuery($"Element `{reader.LocalName}` was unexpected at given location while deserializing type `{Definition.Name}`.");
+                throw new InvalidXRoadQueryException($"Element `{reader.LocalName}` was unexpected at given location while deserializing type `{Definition.Name}`.");
 
             throw XRoadException.UnexpectedElementInQuery(Definition.Name, propertyMap.Definition.Content.SerializedName.LocalName, reader.LocalName);
         }
@@ -105,11 +107,15 @@ namespace XRoadLib.Serialization.Mapping
         private void ValidateRemainingProperties(IEnumerator<IPropertyMap> properties, ContentDefinition content)
         {
             while (properties.MoveNext())
-                if (!properties.Current.Definition.Content.IsOptional)
-                {
-                    var typeName = Definition?.Name ?? ((content.Particle as ArrayItemDefinition)?.Array as PropertyDefinition)?.DeclaringTypeDefinition?.Name;
-                    throw XRoadException.InvalidQuery($"Element `{properties.Current.Definition.Content.SerializedName.LocalName}` is required by type `{typeName}` definition.");
-                }
+            {
+                var propertyMap = properties.Current ?? throw new NullReferenceException();
+                if (propertyMap.Definition.Content.IsOptional)
+                    continue;
+
+                var typeName = Definition?.Name ?? ((content.Particle as ArrayItemDefinition)?.Array as PropertyDefinition)?.DeclaringTypeDefinition?.Name;
+
+                throw new InvalidXRoadQueryException($"Element `{propertyMap.Definition.Content.SerializedName.LocalName}` is required by type `{typeName}` definition.");
+            }
         }
     }
 }