+++ src/XRoadLib/Serialization/XRoadFilterMap.cs
@@ -9,8 +9,8 @@ namespace XRoadLib.Serialization
         private readonly string groupName;
         private readonly ISet<string> enabledProperties = new SortedSet<string>();
 
-        string IXRoadFilterMap.GroupName { get { return groupName; } }
-        ISet<string> IXRoadFilterMap.EnabledProperties { get { return enabledProperties; } }
+        string IXRoadFilterMap.GroupName => groupName;
+        ISet<string> IXRoadFilterMap.EnabledProperties => enabledProperties;
 
         protected XRoadFilterMap(string groupName)
         {
@@ -19,12 +19,11 @@ namespace XRoadLib.Serialization
 
         protected void Enable<TValue>(Expression<Func<T, TValue>> expression)
         {
-            var memberExpression = expression.Body as MemberExpression;
-            if (memberExpression == null)
-                throw new ArgumentException(string.Format("MemberExpression expected, but was {0} ({1}).", expression.Body.GetType().Name, GetType().Name));
+            if (!(expression.Body is MemberExpression memberExpression))
+                throw new ArgumentException($"MemberExpression expected, but was {expression.Body.GetType().Name} ({GetType().Name}).");
 
             if (memberExpression.Expression != expression.Parameters[0])
-                throw new ArgumentException(string.Format("Only parameter members should be used in mapping definition ({0}).", GetType().Name));
+                throw new ArgumentException($"Only parameter members should be used in mapping definition ({GetType().Name}).");
 
             enabledProperties.Add(memberExpression.Member.Name);
         }