+++ src/XRoadLib/Serialization/XRoadMessageReader.cs
@@ -72,7 +72,7 @@ namespace XRoadLib.Serialization
             }
 
             if (target.Header is IXRoadHeader40 xrh4 && xrh4.ProtocolVersion?.Trim() != "4.0")
-                throw XRoadException.InvalidQuery($"Unsupported X-Road v6 protocol version value `{xrh4.ProtocolVersion ?? string.Empty}`.");
+                throw new InvalidXRoadQueryException($"Unsupported X-Road v6 protocol version value `{xrh4.ProtocolVersion ?? string.Empty}`.");
 
             if (target.IsMultipartContainer)
                 target.BinaryMode = BinaryMode.Attachment;
@@ -87,7 +87,7 @@ namespace XRoadLib.Serialization
                 return;
 
             if (!Equals(target.RootElementName.LocalName, target.Header.Service.ServiceCode))
-                throw XRoadException.InvalidQuery($"Teenuse nimi `{target.Header.Service.ServiceCode}` ei ole vastavuses päringu sisuga `{target.RootElementName}`.");
+                throw new InvalidXRoadQueryException($"X-Road operation name `{target.Header.Service.ServiceCode}` does not match request wrapper element name `{target.RootElementName}`.");
         }
 
         public void Dispose()
@@ -372,7 +372,7 @@ namespace XRoadLib.Serialization
         private IServiceManager DetectServiceManager(XmlReader reader)
         {
             if (!reader.MoveToElement(0, "Envelope", NamespaceConstants.SOAP_ENV))
-                throw XRoadException.InvalidQuery("Päringus puudub SOAP-ENV:Envelope element.");
+                throw new InvalidXRoadQueryException("X-Road SOAP request is missing SOAP-ENV:Envelope element.");
 
             return serviceManagers.SingleOrDefault(p => p.IsDefinedByEnvelope(reader));
         }